Refactor Trip Details Page to use Client Component and Implement Static Generation
- Removed client-side logic from the TripDetailsPage component and created a new TripDetailsClient component for better separation of concerns.
- Implemented static generation for trip details by adding generateStaticParams function to fetch trip IDs at build time.
- Updated next.config.mjs to remove standalone output configuration.
